ai_hackthon_backend/requirements.txt
Django==4.2.29
djangorestframework==3.15.2
drf-yasg==1.21.8
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.157.0
